backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
PyJWT==2.10.1
google-cloud-storage==3.2.0
google-cloud-firestore==2.21.0
google-genai==1.29.0
pypdf==5.9.0
Pillow==11.3.0
numpy==2.3.3
httpx==0.28.1
email-validator==2.2.0
firebase-admin==7.1.0
